Backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
distlib==0.3.8
django==4.2.9; python_version >= '3.8'
django-rest-framework==0.1.0
djangorestframework==3.14.0; python_version >= '3.6'
filelock==3.13.1; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
mysqlclient==2.2.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pillow==10.2.0; python_version >= '3.8'
pipenv==2023.11.17; python_version >= '3.8'
platformdirs==4.1.0; python_version >= '3.8'
pluggy==1.3.0; python_version >= '3.8'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tox==3.28.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
tox-pipenv==1.10.1
urllib3==2.1.0; python_version >= '3.8'
virtualenv==20.25.0; python_version >= '3.7'
